fixed simulation preprocessing and modulation problems

+++ modulator.py
@@ -1,8 +1,10 @@
+import numpy as np
+
 def modulate(value, max = 255, newMax = 15):
     valueN = value * (newMax/ max) # Linearly normzalized value
     valueN = int(round(valueN, 0)) # round to nearest integer
     value = valueN * (max/newMax) # Modulated value
-    return int(value)
+    return np.uint8(value)
 
 
 def main():
